+++ lib/sedna/algorithms/aggregation/__init__.py
@@ -12,45 +12,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Aggregation algorithms"""
-
-import abc
-from copy import deepcopy
-
-import numpy as np
-
-from sedna.common.class_factory import ClassFactory, ClassType
-
-__all__ = ('FedAvg',)
-
-
-class BaseAggregation(metaclass=abc.ABCMeta):
-    def __init__(self):
-        self.total_size = 0
-        self.weights = None
-
-    @abc.abstractmethod
-    def aggregate(self, weights, size=0):
-        """
-        Aggregation
-        :param weights: deep learning weight
-        :param size: numbers of sample in each loop
-        """
-
-
-@ClassFactory.register(ClassType.FL_AGG)
-class FedAvg(BaseAggregation, abc.ABC):
-    """Federated averaging algorithm"""
-
-    def aggregate(self, weights, size=0):
-        total_sample = self.total_size + size
-        if not total_sample:
-            return self.weights
-        updates = []
-        for inx, weight in enumerate(weights):
-            old_weight = self.weights[inx]
-            row_weight = ((np.array(weight) - old_weight) *
-                          (size / total_sample) + old_weight)
-            updates.append(row_weight)
-        self.weights = deepcopy(updates)
-        return updates
+from .aggregation import *
