+++ lib/sedna/algorithms/unseen_task_detect/__init__.py
@@ -12,60 +12,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Unseen Task detect Algorithms for Lifelong Learning"""
-
-import abc
-from typing import List
-
-import numpy as np
-
-from sedna.algorithms.multi_task_learning.task_jobs.artifact import Task
-from sedna.common.class_factory import ClassFactory, ClassType
-
-__all__ = ('ModelProbeFilter', 'TaskAttrFilter')
-
-
-class BaseFilter(metaclass=abc.ABCMeta):
-    """The base class to define unified interface."""
-
-    def __call__(self, task: Task = None):
-        """predict function, and it must be implemented by
-        different methods class.
-
-        :param task: inference task
-        :return: `True` means unseen task, `False` means not an unseen task.
-        """
-        raise NotImplementedError
-
-
-@ClassFactory.register(ClassType.UTD, alias="ModelProbe")
-class ModelProbeFilter(BaseFilter, abc.ABC):
-    def __init__(self):
-        pass
-
-    def __call__(self, tasks: List[Task] = None, threshold=0.5, **kwargs):
-        all_proba = []
-        for task in tasks:
-            sample = task.samples
-            model = task.model
-            if hasattr(model, "predict_proba"):
-                proba = model.predict_proba(sample)
-                all_proba.append(np.max(proba))
-        return np.mean(all_proba) > threshold if all_proba else True
-
-
-@ClassFactory.register(ClassType.UTD, alias="TaskAttr")
-class TaskAttrFilter(BaseFilter, abc.ABC):
-    def __init__(self):
-        pass
-
-    def __call__(self, tasks: List[Task] = None, **kwargs):
-        for task in tasks:
-            model_attr = list(map(list, task.model.meta_attr))
-            sample_attr = list(map(list, task.samples.meta_attr))
-
-            if not (model_attr and sample_attr):
-                continue
-            if list(model_attr) == list(sample_attr):
-                return False
-        return True
+from .unseen_task_detect import *
